+++ bcdi/postprocessing/postprocessing_utils.py
@@ -2209,7 +2209,9 @@ def tukey_window(shape, alpha=np.array([0.5, 0.5, 0.5])):
     return tukey3
 
 
-def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
+def unwrap(
+    obj, support_threshold, seed=0, skip_unwrap: bool = False, debugging=True, **kwargs
+):
     """
     Unwrap the phase of a complex object.
 
@@ -2220,6 +2222,8 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
     :param support_threshold: relative threshold used to define a support from abs(obj)
     :param seed: int, random seed. Use always the same value if you want a
      deterministic behavior.
+    :param skip_unwrap: True to apply only the support_threshold to the phase but skip
+     the unwrapping part.
     :param debugging: set to True to see plots
     :param kwargs:
 
@@ -2247,28 +2251,32 @@ def unwrap(obj, support_threshold, seed=0, debugging=True, **kwargs):
 
     ndim = obj.ndim
     unwrap_support = np.ones(obj.shape, dtype=int)
-    unwrap_support[
-        abs(obj) > support_threshold * abs(obj).max()
-    ] = 0  # 0 is a valid entry for ma.masked_array
-    phase_wrapped = ma.masked_array(np.angle(obj), mask=unwrap_support)
+    unwrap_support[abs(obj) > support_threshold * abs(obj).max()] = 0
+    # 0 is a valid entry for ma.masked_array
+    phase_wrapped: np.ndarray = ma.masked_array(np.angle(obj), mask=unwrap_support)
 
+    plot_title = "applying support threshold\n" if skip_unwrap else "unwrapping"
     if debugging and ndim == 3:
         gu.multislices_plot(
             phase_wrapped.data,
             plot_colorbar=True,
-            title="Object before unwrapping",
+            title=f"Before {plot_title}",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
             cmap=kwargs.get("cmap", "turbo"),
         )
 
-    phase_unwrapped = unwrap_phase(phase_wrapped, wrap_around=False, seed=seed).data
+    if skip_unwrap:
+        phase_unwrapped = phase_wrapped
+    else:
+        phase_unwrapped = unwrap_phase(phase_wrapped, wrap_around=False, seed=seed).data
+
     phase_unwrapped[np.nonzero(unwrap_support)] = 0
     if debugging and ndim == 3:
         gu.multislices_plot(
             phase_unwrapped,
             plot_colorbar=True,
-            title="Object after unwrapping",
+            title=f"After {plot_title}",
             reciprocal_space=reciprocal_space,
             is_orthogonal=is_orthogonal,
             cmap=kwargs.get("cmap", "turbo"),